+++ lib/parse/cmdline.py
@@ -849,6 +849,12 @@ def parse_arguments(arguments: list[str] | None = None) -> Values:
         dest="crawl",
         help="Crawl for new paths in responses"
     )
+    advanced.add_option(
+        "--find-backup",
+        action="store_true",
+        dest="find_backup",
+        help="Look for backups of discovered files",
+    )
 
     # View Settings
     view = OptionGroup(parser, "View Settings")